UserContext: keep value from GET /me
We do have this information in our redux store, but it is annoying to always do
checks on that RemoteData. Since we require the user data do be fetched when we
start the application, we can structure it in such a way that we can be certain
that we always have the user data available. We do this by putting it in its own
context that components can use to access this information in a safe way.