Simplify handling of API errors
Instead of discarding much of the information in the error we place more
responsibility and flexibility on the views. The previous implementation would
also lie a bit about the error code, as our enum hid information by saying that
it could only be Unauthorized, NotAuthenticated or Unknown wheras in reality the
field would hold the actual response code from the server.