Force usage of node-forge 0.10.0
It's a dependency of parcel, and as they are unwilling/unable update it
on their end we have to force it.
See https://github.com/parcel-bundler/parcel/issues/5145.

Having to add the fix both in package.json and server/package.json
also highlights that our setup is not that great. We have parcel-bundler
as a dependency in package.json for building the frontend and in
server/package.json for running the dev version. Ideally we should only
have it as a dependency one of these places.